pyproject.toml
[project]
name = "sentinel-triage-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "fastmcp>=3.2.4",
    "langchain-google-genai>=4.2.2",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.9",
    "msal>=1.36.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
